# Conflicts:
#	app/src/main/AndroidManifest.xml

+++ app/src/main/java/com/wd/tech/model/bean/information/DetailsBean.java
@@ -208,7 +208,7 @@ public class DetailsBean {
             return yuanCost;
         }
 
-        public void setYuanCost(int yuanCost) {
+        public void setYuanCost(double yuanCost) {
             this.yuanCost = yuanCost;
         }
